keep Modals component instead of renaming to ModalStack
