+++ astro.config.mjs
@@ -31,7 +31,7 @@ export default defineConfig({
             // Each item here is one entry in the navigation menu.
             { label: 'Getting Started', slug: 'usage/getting-started' },
             { label: 'Modal Components', slug: 'usage/modal-components' },
-            { label: '<ModalStack />', slug: 'usage/modal-stack' },
+            { label: '<Modals />', slug: 'usage/modal-stack' },
             { label: 'API Reference', slug: 'usage/api' },
             { label: 'v2 Migration', slug: 'usage/v2-migration' }
           ]
