+++ docs/content/docs/usage/modal-stack.mdx
@@ -1,5 +1,5 @@
 ---
-title: <ModalStack />
+title: <Modals />
 layout: ../../layouts/CustomRenderingLayout.astro
 ---
 
@@ -7,10 +7,10 @@ Responsible for rendering the current stack of modals.
 
 ## Props
 
-ModalStack can receive the following props (which are currently all just snippets):
+Modals can receive the following props (which are currently all just snippets):
 
 ```ts
-interface ModalStackProps {
+interface ModalsProps {
   backdrop?: Snippet<[modals: Modals]>
   modal?: Snippet<
     [
@@ -32,11 +32,11 @@ interface ModalStackProps {
 Renders when at least 1 modal has been opened.
 
 ```svelte
-<ModalStack>
+<Modals>
   {#snippet backdrop(modals)}
     <div class="backdrop" onclick={() => modals.close()}></div>
   {/snippet}
-</ModalStack>
+</Modals>
 ```
 
 ### modal
@@ -44,11 +44,11 @@ Renders when at least 1 modal has been opened.
 Renders each modal component in the stack
 
 ```svelte
-<ModalStack>
+<Modals>
   {#snippet modal(modal, modals)}
     <modal.component {...modal.props} />
   {/snippet}
-</ModalStack>
+</Modals>
 ```
 
 ### loading
@@ -56,11 +56,11 @@ Renders each modal component in the stack
 Renders when a modal has been opened via dynamic import and is still loading.
 
 ```svelte
-<ModalStack>
+<Modals>
   {#snippet loading()}
     <LoadingSpinner />
   {/snippet}
-</ModalStack>
+</Modals>
 ```
 
 ## Advanced Usage
@@ -72,10 +72,10 @@ to always show each opened modal.
 
 ```svelte
 <script>
-  import { ModalStack, modals } from 'svelte-modals'
+  import { Modals, modals } from 'svelte-modals'
 </script>
 
-<ModalStack>
+<Modals>
   {#snippet backdrop({ close })}
     <div class="backdrop" onclick={() => close()}></div>
   {/snippet}
@@ -86,7 +86,7 @@ to always show each opened modal.
       isOpen={modals.stack.length > 0} 
     />
   {/snippet}
-</ModalStack>
+</Modals>
 ```
 
 ```svelte live layout="../../layouts/code/NoCode.astro"
