+++ docs/content/layouts/AllModals.svelte
@@ -1,8 +1,8 @@
 <script>
-  import { ModalStack } from 'svelte-modals'
+  import { Modals } from 'svelte-modals'
 </script>
 
-<ModalStack>
+<Modals>
   {#snippet backdrop({ close })}
     <div class="backdrop" onclick={() => close()}></div>
   {/snippet}
@@ -10,4 +10,4 @@
   {#snippet modal(modal, modals)}
     <modal.component {...modal.props} isOpen={modals.stack.length > 0} />
   {/snippet}
-</ModalStack>
+</Modals>
