+++ docs/content/layouts/AnimatedBackdropLayout.astro
@@ -1,6 +1,6 @@
 ---
-import ModalStack  from './ModalStack.svelte'
+import Modals  from './Modals.svelte'
 ---
 
-<ModalStack client:load fade />
+<Modals client:load fade />
 <slot />
