+++ docs/content/layouts/BaseLayout.astro
@@ -1,6 +1,6 @@
 ---
-import ModalStack  from './ModalStack.svelte'
+import Modals  from './Modals.svelte'
 ---
 
-<ModalStack client:load />
+<Modals client:load />
 <slot />
