+++ docs/content/layouts/CustomRenderingLayout.astro
@@ -1,6 +1,6 @@
 ---
-import ModalStack  from './ShowAllModalStack.svelte'
+import Modals  from './AllModals.svelte'
 ---
 
-<ModalStack client:load />
+<Modals client:load />
 <slot />
