--- docs/content/layouts/LegacyBaseLayout.astro
@@ -1,6 +1,0 @@
----
-import LegacyModalStack  from './LegacyModalStack.svelte'
----
-
-<LegacyModalStack client:load />
-<slot />
